backend/requirements/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-stubs==5.1.1
django-stubs-ext==5.1.1
django-timescaledb==0.2.13
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.2
Faker==37.1.0
gunicorn==23.0.0
idna==3.10
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
ruff==0.9.2
sqlparse==0.5.3
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
typing_extensions==4.11.0
tzdata==2025.2
urllib3==2.3.0
